### 2020-08-27 d4f4009
Refactor identification of segments.
