+++ biopeaks/tests/test_gui.py
@@ -366,11 +366,11 @@ ecg_batch_os = {"modality": "ECG",
                 "sigfnames": ["OSmontage1A.txt", "OSmontage1J.txt",
                               "OSmontage2A.txt", "OSmontage2J.txt",
                               "OSmontage3A.txt", "OSmontage3J.txt"],
-                "peaksums": [3808244, 3412308, 2645824, 3523449, 3611836,
-                             3457936],
-                "stats": [(0.7950, 76.1123), (0.7288, 83.1468),
-                          (0.7894, 76.8911), (0.7402, 81.7864),
-                          (0.7856, 76.9153), (0.7235, 83.6060)],
+                "peaksums": [3828709, 3434713, 2666196, 3546319, 3611836,
+                             3480340],
+                "stats": [(0.7916, 76.3624), (0.7285, 83.1771),
+                          (0.7889, 76.9233), (0.7402, 81.7879),
+                          (0.7856, 76.9153), (0.7234, 83.6239)],
                 "correctpeaks": False}
 
 ecg_batch_custom = {"modality": "ECG",
@@ -381,11 +381,11 @@ ecg_batch_custom = {"modality": "ECG",
                     "sigfnames": ["OSmontage1A.txt", "OSmontage1J.txt",
                                   "OSmontage2A.txt", "OSmontage2J.txt",
                                   "OSmontage3A.txt", "OSmontage3J.txt"],
-                    "peaksums": [3808244, 3412308, 2645824, 3523449, 3611836,
-                                 3457936],
-                    "stats": [(0.7950, 76.1123), (0.7288, 83.1468),
-                              (0.7894, 76.8911), (0.7402, 81.7864),
-                              (0.7856, 76.9153), (0.7235, 83.6060)],
+                    "peaksums": [3828709, 3434713, 2666196, 3546319, 3611836,
+                                 3480340],
+                    "stats": [(0.7916, 76.3624), (0.7285, 83.1771),
+                              (0.7889, 76.9233), (0.7402, 81.7879),
+                              (0.7856, 76.9153), (0.7234, 83.6239)],
                     "correctpeaks": False}
 
 ecg_batch_autocorrect = {"modality": "ECG",
@@ -395,11 +395,11 @@ ecg_batch_autocorrect = {"modality": "ECG",
                          "sigfnames": ["OSmontage1A.txt", "OSmontage1J.txt",
                                        "OSmontage2A.txt", "OSmontage2J.txt",
                                        "OSmontage3A.txt", "OSmontage3J.txt"],
-                         "peaksums": [3808199, 3394481, 2626449, 3511241,
-                                      3611833, 3457931],
-                         "stats": [(0.7944, 76.0973), (0.7308, 82.8572),
-                                   (0.7934, 76.2351), (0.7418, 81.5011),
-                                   (0.7856, 76.9152), (0.7235, 83.5973)],
+                         "peaksums": [3828678, 3416886, 2644580, 3534109,
+                                      3611833, 3480341],
+                         "stats": [(0.7914, 76.3427), (0.7305, 82.8876),
+                                   (0.7942, 75.954), (0.7418, 81.5025),
+                                   (0.7856, 76.9152), (0.7233, 83.6178)],
                          "correctpeaks": True}
 
 
