### 2020-08-31 #2621 @whedon
[REVIEW]: biopeaks: a graphical user interface for feature extraction from heart- and breathing biosignals
**Submitting author:** @JanCBrammer (<a href="http://orcid.org/0000-0002-7664-3753">Jan C. Brammer</a>)
**Repository:** <a href="https://github.com/JanCBrammer/biopeaks" target ="_blank">https://github.com/JanCBrammer/biopeaks</a>
**Version:** v1.4.1

- [x] **Functionality:** Have the functional claims of the software been confirmed? https://github.com/JanCBrammer/biopeaks/issues/14#issue-694096081

- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. https://github.com/JanCBrammer/biopeaks/pull/13
- [x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). https://github.com/JanCBrammer/biopeaks/issues/12
- [x] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? https://github.com/JanCBrammer/biopeaks/issues/11 https://github.com/JanCBrammer/biopeaks/issues/10
- [x] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified? https://github.com/JanCBrammer/biopeaks/issues/8
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support https://github.com/JanCBrammer/biopeaks/issues/9

- [x] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)? https://github.com/JanCBrammer/biopeaks/pull/6
