### 2020-08-31 #2621 @oliviaguest
👋👋👋

Hey @JanCBrammer @TomDonoghue, @sappelhoff, @jamesheathers! Welcome to where the review itself will take place. Please make sure to read the instructions above.

For any and all things worthy of discussion or comment, use this issue right here — so drop comments or questions for me, the author, etc., here. For any very code-specific things please feel free to start an issue on the repo of the code itself (if appropriate!) and link back to it from here. For an example of how this process plays out feel free to skim previous reviews, such as: https://github.com/openjournals/joss-reviews/issues/2285 and https://github.com/openjournals/joss-reviews/issues/2348. ☺️
