### 2020-09-02 JanCBrammer/biopeaks#4 @sappelhoff
potentially add .mailmap for consistent commiter identities
Hi @JanCBrammer, I am one of the reviewers of your submission at JOSS https://github.com/openjournals/joss-reviews/issues/2621 and over the next weeks I'll dive into the repo and ... review it :-)

I'll try to make small and targeted issues whenever I can, this is the first one:

When trying to find which people contributed to this code, I ran `git shortlog --summary --email` and got the following output:

```Text
     7  Jan C. Brammer <j.brammer@psych.ru.nl>
     1  Jan C. Brammer <jan.c.brammer@gmail.com>
   242  JohnDoenut <j.brammer@psych.ru.nl>
   137  JohnDoenut <jan.c.brammer@gmail.com>
```

I can see that it's all you, but you might want to consider adding a [`.mailmap` file](https://git-scm.com/docs/git-check-mailmap#_mapping_authors):

> it is used to map author and committer names and email addresses to canonical real names and email addresses

so that you can merge your 4 identities into one reasonable identity (of your choice).

For example, see this `.mailmap` --> https://github.com/mne-tools/mne-bids/blob/master/.mailmap

Edit: To clarify the goal: if there is only one author, then running `git shortlog --summary --email` should also only yield one output (not 4 as in this case) ... and that goal can be achieved by `.mailmap`
