### 2020-09-03 JanCBrammer/biopeaks#4 @JanCBrammer
Hi @sappelhoff, thanks for reviewing the repositoy and thanks for the useful pointer!

> if there is only one author, then running `git shortlog --summary --email` should also only yield one output

Works like a charm.
```
390  Jan C. Brammer <jan.c.brammer@gmail.com>
```
