### 2020-09-05 JanCBrammer/biopeaks#5 @sappelhoff
Adding .zenodo.json?
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

**EDIT:** I just saw that you already added the zenodo badge. I probably oversaw it because I was expecting it at the top of the README next to the testing badge. Feel free to ignore my comment on that badge, However, my comment on the `.zenodo.json` is still valid.

---

Using a Google search I saw that you have stored a record of biopeaks on Zenodo -> https://zenodo.org/record/3700054

(probably via the GitHub integration, but I cannot check right now because Zenodo seems to be down)

I think that's great. Here are two suggestions:

1. add the zenodo badge to your repository README (alternatively, don't add it, and add the JOSS badge instead) --> the reasoning is to point users to an archived version of the code that will persist even if GitHub is deleted someday. Once (if) your paper is published in JOSS, the JOSS badge would be sufficient, because the JOSS paper will also contain a link to the Zenodo archive.
1. (only if you are in fact using the Zenodo-GitHub integration, which I couldn't verify at the moment of writing) --> add a `.zenodo.json` file to the root of your project. These files are not very well documented (unfortunately ... see https://github.com/zenodo/zenodo/issues/1606), but they allow you to specify metadata for the Zenodo record. As an example, see: https://github.com/mne-tools/mne-bids/blob/master/.zenodo.json
