+++ paper/paper.bib
@@ -4,7 +4,7 @@
   title = {{BioSPPy}: Biosignal Processing in {Python}},
   year = {2015--},
   url = "https://github.com/PIA-Group/BioSPPy/",
-  note = {[Online; accessed <today>]}
+  note = {[Online; accessed 23 August 2020]}
 }
 
 
