### 2020-09-05 2b7c5b0
fix pandas citation via https://pandas.pydata.org/about/citing.html
