+++ paper/paper.bib
@@ -71,11 +71,12 @@
 
 @inproceedings{pandas,
 	title = {Data Structures for Statistical Computing in Python},
-	pages = {51 -- 56},
+	pages = {56 -- 61},
 	booktitle = {Proceedings of the 9th Python in Science Conference},
 	author = {{McKinney}, Wes},
-	editor = {Walt, Stéfan van der and Millman, Jarrod},
-	date = {2010}
+	editor = {van der Walt, Stéfan and Millman, Jarrod},
+	date = {2010},
+	doi = {10.25080/Majora-92bf1922-00a}
 }
 
 
