### 2020-09-05 923a39f
fix doi by removing https prefix
