### 2020-09-05 JanCBrammer/biopeaks#7 @sappelhoff
Suggestion: Add classifiers, long_description, and more links to setup.py
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

Looking at your package on [pypi](https://pypi.org/project/biopeaks/), I see that a description is missing there:

![image](https://user-images.githubusercontent.com/9084751/92302175-50d61280-ef6a-11ea-89b8-f0a70d4dcd99.png)

Personally, I prefer seeing a description on pypi -> and usually it's easy to re-use the README from GitHub and have it rendered on pypi as well. For example, see [these lines in another package](https://github.com/mne-tools/mne-bids/blob/f28caec7b58cb9ca406cd8b1a3152879f3d85729/setup.py#L38-L39). You need to populate `long_description` in `setup.py`

Next to the description, I would advise you to add `classifiers` to your `setup.py` (https://pypi.org/classifiers/) --> metadata makes it easier for users (and other software) to find your package, or to filter packages according to their needs.

Finally, I want to suggest to you to add more links to your `setup.py`. You can use the following syntax ([link to example](https://github.com/mne-tools/mne-bids/blob/f28caec7b58cb9ca406cd8b1a3152879f3d85729/setup.py#L61-L64)):

```
          project_urls={
              'Documentation': 'https://mne.tools/mne-bids',
              'Bug Reports': 'https://github.com/mne-tools/mne-bids/issues',
              'Source': 'https://github.com/mne-tools/mne-bids',
          },
```

which then renders on pypi:

![image](https://user-images.githubusercontent.com/9084751/92302246-e8d3fc00-ef6a-11ea-8fdf-c8fd8cee7c55.png)

PS: You can also edit the GitHub "About" section to add the link to your documentation:

![image](https://user-images.githubusercontent.com/9084751/92302275-38b2c300-ef6b-11ea-8f62-42026a58d896.png)

then users won't have to go through the README to find the link to the docs.
