### 2020-09-05 JanCBrammer/biopeaks#8 @sappelhoff
Add code coverage to tests
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

It would be great to add a tool that measures what proportion of your code is actually covered by the tests.

It's helpful to:
- give users a better impression of how far they can "trust" the tests
- give the developers a better idea of what aspects of the code are tested, and what aspects are not tested

This would be quite easy to implement, e.g., using pytest extensions and [codecov](https://codecov.io/) as in this repository: https://github.com/sappelhoff/pyprep

- you'd need `pytest-cov` (https://github.com/sappelhoff/pyprep/blob/da8e3a64cac7759e553d9dc219a88edc13f3b6cb/requirements-dev.txt#L10)
- you'd need to extend your GitHub Actions workflow (example: https://github.com/sappelhoff/pyprep/blob/da8e3a64cac7759e553d9dc219a88edc13f3b6cb/.github/workflows/python_tests.yml#L56-L60)

Finally, you could add a coverage badge next to your tests:

![image](https://user-images.githubusercontent.com/9084751/92302427-76641b80-ef6c-11ea-8aec-f30f020e9f34.png)
