### 2020-09-05 JanCBrammer/biopeaks#9 @sappelhoff
Making community and contributing guidelines more visible
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

At the moment you have a [contributor's guide in the documentation](https://jancbrammer.github.io/biopeaks/contributor_guide.html). That's cool and I have some suggestions on how to make this more visible:

# New sections / re-ordering of README that is displayed on GitHub.
- most users will see this README as the first thing
- a new section `# Documentation` would be nice --> the heading will make it immediately obvious where people have to click
    - currently the link to the docs is buried in "general info"
    - perhaps also worth thinking about which type of documentation you publish --> is it the "stable" version or "latest" (i.e., unreleased changes on `master`) or both?
- another new section `# Contributing` would be good with a brief sentence that contributions (of kind XYZ) are welcome ... and then provide a direct link to the contributors guide in the docs

# Code of Conduct
- currently there have not been many contributions from other people, but as soon as that starts to happen, you'd probably be better off with a code of conduct
- It's relatively low cost to add such a code and it can help if conflict ever arises
- see for example: https://www.contributor-covenant.org/
- GitHub organizations can also have a "default" code of conduct that gets "transmitted" to all repositories of that organization that do not yet have an own "code of conduct". This happens via a `.github` repository (see e.g., https://github.com/bids-standard/.github/) --> might be worth checking whether this is also possible for GitHub *users* (as opposed to organizations)
- docs for that: https://docs.github.com/en/github/building-a-strong-community/creating-a-default-community-health-file
- else, just add your own `CODE_OF_CONDUCT.md` to the root of this repo
