### 2020-09-05 JanCBrammer/biopeaks#10 @sappelhoff
Suggestion: Add docs on how docs are being built
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

I think it'd be helpful for potential contributors (or your future self) to have a short paragraph with information on how the documentation is being build and deployed. A good place for this documentation would be in the contributors guide.
