### 2020-09-05 JanCBrammer/biopeaks#11 @sappelhoff
Discussion: Adopting a consistent style for docstrings
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

Looking at the code base, I find functions / methods / classes that are sometimes not documented, and sometimes extensively documented but in a seemingly unstructured way. Sometimes I also see a documentation about function parameters such as here:

https://github.com/JanCBrammer/biopeaks/blob/89fe7c8c80de29549951e6f7fed75fb21506903a/biopeaks/benchmarks/benchmark_utils.py#L17-L19

but more often than not, the parameters are not documented.

**I would strongly recommend to adopt a consistent docstr style** (such as [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html)) **and apply this to the codebase**. The style can also be enforced with `flake8` extensions such as  [`flake8-docstrings`](https://pypi.org/project/flake8-docstrings/).

Consistent, extensive (but concise) docstrings make life easier for contributors, reviewers, and your future self ... and perhaps also for users --> but for that last part I am not entirely sure, because this a GUI package after all, and users are not expected to work with an API.

That latter sentence is the reason why I label this issue as "Discussion" ... I can imagine that other people have different opinions about this point.
