### 2020-09-05 JanCBrammer/biopeaks#10 @sappelhoff
another note: when adding docs on how to build docs, perhaps it'd be worth adding also docs for contributors on how to install a "development version" of the repository (which is required for making PRs)
