### 2020-09-05 JanCBrammer/biopeaks#12 @sappelhoff
Make use of a dedicated example file in "getting-started" section in docs
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

I was just trying to follow the [getting started](https://jancbrammer.github.io/biopeaks/user_guide.html#getting-started) section in the docs, and noticed that the "examplary workflow on a single file" stays relatively abstract.

I think it'd be great to add a dedicated example file to the repository and make use of that file in that section of the docs: Instead of describing the functions as currently done, describe single steps and expect users to follow along.

For that to work, example files should ideally be shipped with the installation and/or the docs should explain how to easily get these files.

There should also be a practical example of reading a file in "custom" format, like [here](https://jancbrammer.github.io/biopeaks/user_guide.html#plain-text-files) but with a real file that users can open and look at.
