### 2020-09-05 JanCBrammer/biopeaks#11 @sappelhoff
Documentation with docstrings for me also includes writing an "introductory docstring" for **each** public module (almost every .py file).

For example the `benchmark_ECG_stream.py` file looks like this:

```Python
from biopeaks.heart import ecg_peaks
from benchmark_utils import BenchmarkDetectorGUDB

pipeline = BenchmarkDetectorGUDB(ecg_peaks, 1)
pipeline.benchmark_records("jogging", channel="cs_V2_V1", annotation="annotation_cs")

```

But I  think it should look like the following.

(Note 1: and Python docstring convention also thinks that way :slightly_smiling_face: see pep [257](https://www.python.org/dev/peps/pep-0257/) and the D100 error flake8 gives you when you don't write a module level docstring)

(Note 2: the *content* of the docstr is made up, just as an example)

```Python
"""Benchmark the ECG algorithm against the database.

Here we stream data from XYZ and feed it through our algorithm.
Important resources to check are: http.I.like.it

"""
from biopeaks.heart import ecg_peaks
from benchmark_utils import BenchmarkDetectorGUDB

pipeline = BenchmarkDetectorGUDB(ecg_peaks, 1)
pipeline.benchmark_records("jogging", channel="cs_V2_V1", annotation="annotation_cs")

```
