### 2020-09-05 JanCBrammer/biopeaks#14 @sappelhoff
Question/suggestions about benchmark files
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

In the [tests](https://jancbrammer.github.io/biopeaks/tests.html) section of the docs you explain nicely how one benchmark from `biopeaks/benachmarks` can be run.

For the other benchmarks I feel like some documentation would come in handy (related to https://github.com/JanCBrammer/biopeaks/issues/11#issuecomment-687616629)

For example I think that I as a user should set these two lines from `benchmark_PPG.py` to the path where I download the `Capnobase IEEE TBME benchmark dataset` to?

```
record_dir = r"C:\Users\JohnDoe\surfdrive\Beta\example_data\PPG\signal"
annotation_dir = r"C:\Users\JohnDoe\surfdrive\Beta\example_data\PPG\annotations"
records = os.listdir(record_dir)
```

and I assume that `benchmark_ECG_local.py` can be used when the GUDB data has been downloaded instead of streaming it via `benchmark_ECG_stream.py`? --> all of this should be cleared up using module level docstrings (see link to doc comment above) and smaller inline comments.

Another note: "uncommented" code in scripts like here ...

https://github.com/JanCBrammer/biopeaks/blob/89fe7c8c80de29549951e6f7fed75fb21506903a/biopeaks/benchmarks/benchmark_ECG_local.py#L55-L58

... could be improved with an if / else clause and a simple setting of a boolean value at the top of the script. that way of coding is often easier understood and leads to less errors.

---

Suggestion in case you didn't know: Documenting code with sphinx and [sphinx-gallery](https://sphinx-gallery.github.io/stable/index.html) allows you to integrate such "benchmarks" as examples that get built automatically upon building your documentation --> that way users can inspect the code and all of its results online from their browser, instead of having to run it themselves.

You can see that in action here: https://mne.tools/mne-bids/stable/auto_examples/convert_eeg_to_bids.html#sphx-glr-auto-examples-convert-eeg-to-bids-py
