### 2020-09-05 JanCBrammer/biopeaks#14 @sappelhoff
One more question about `benchmark_PPG.py` --> according to the docs, I should first download the data from http://www.capnobase.org/index.php?id=857

I did that (specifically: `TBME2013-PPGRR-Benchmark_R3.zip`), and the data looks like the following. I cannot find "signal" and "annotations" as required in `benchmark_PPG.py`. where am I going wrong?

```
TBME2013-PPGRR-Benchmark_R3
├── data
│   ├── 0009_8min.mat
│   ├── 0015_8min.mat
│   ├── 0016_8min.mat
│   ├── 0018_8min.mat
│   ├── 0023_8min.mat
│   ├── 0028_8min.mat
│   ├── 0029_8min.mat
│   ├── 0030_8min.mat
│   ├── 0031_8min.mat
│   ├── 0032_8min.mat
│   ├── 0035_8min.mat
│   ├── 0038_8min.mat
│   ├── 0103_8min.mat
│   ├── 0104_8min.mat
│   ├── 0105_8min.mat
│   ├── 0115_8min.mat
│   ├── 0121_8min.mat
│   ├── 0122_8min.mat
│   ├── 0123_8min.mat
│   ├── 0125_8min.mat
│   ├── 0127_8min.mat
│   ├── 0128_8min.mat
│   ├── 0133_8min.mat
│   ├── 0134_8min.mat
│   ├── 0142_8min.mat
│   ├── 0147_8min.mat
│   ├── 0148_8min.mat
│   ├── 0149_8min.mat
│   ├── 0150_8min.mat
│   ├── 0309_8min.mat
│   ├── 0311_8min.mat
│   ├── 0312_8min.mat
│   ├── 0313_8min.mat
│   ├── 0322_8min.mat
│   ├── 0325_8min.mat
│   ├── 0328_8min.mat
│   ├── 0329_8min.mat
│   ├── 0330_8min.mat
│   ├── 0331_8min.mat
│   ├── 0332_8min.mat
│   ├── 0333_8min.mat
│   └── 0370_8min.mat
└── README.txt

1 directory, 43 files

```
