### 2020-09-10 #2621 @JanCBrammer
@sappelhoff, thanks for your thorough and helpful review so far, much appreciated! I'm offline next week (week 38), so I won't get around to address your questions and suggestions until the week after the next (week 39). Looking forward to it :)
