### 2020-09-22 41d7a6f
Run PPG benchmark directly on the Capnobase download (no need to unpack data first).
