+++ docs/tests.md
@@ -17,6 +17,8 @@ pytest -v
 
 
 ## Extrema detection benchmarks
+
+### ECG
 To validate the performance of the ECG peak detector `heart.ecg_peaks()`, please install the [wfdb](https://github.com/MIT-LCP/wfdb-python) and [aiohttp](https://github.com/aio-libs/aiohttp):
 ```
 conda install -c conda-forge wfdb
@@ -24,8 +26,17 @@ conda install -c conda-forge aiohttp
 ```
 
 You can then run the `benchmark_ECG_stream` script in the `benchmarks` folder. The script streams ECG and annotation files from the [Glasgow University Database (GUDB)](http://researchdata.gla.ac.uk/716/).
-You can select an experiment, ECG channel, and annotation file.
+You can select an experiment, ECG channel, and annotation file (for details have a look at the docstrings of `BenchmarkDetectorGUDB.benchmark_records()` in `benchmarks\benchmark_utils`).
+
+Alternatively, you can download the GUDB and run the `benchmark_ECG_local` script in the `benchmarks` folder. In the script, replace the `data_dir` with your local directory (see comments in the script).
+
+### PPG
 
 To validate the performance of the PPG peak detector `heart.ppg_peaks()`
-please download the [Capnobase IEEE TBME benchmark dataset](http://www.capnobase.org/index.php?id=857).
-After extracting the PPG signals and peak annotations you can run the `benchmark_PPG` script in the `benchmarks` folder.
+please download the [Capnobase IEEE TBME benchmark dataset](http://www.capnobase.org/index.php?id=857) and install [wfdb](https://github.com/MIT-LCP/wfdb-python) and [h5py](https://www.h5py.org/):
+```
+conda install -c conda-forge wfdb
+conda install -c conda-forge h5py
+```
+
+You can then run the `benchmark_PPG_local` script in the `benchmarks` folder. In the script, replace the `data_dir` with your local directory (see comments in the script).
