### 2020-09-22 8abe469
Use pathlib and make script more consistent with PPG benchmark script.
