+++ biopeaks/benchmarks/benchmark_ECG_local.py
@@ -1,35 +1,27 @@
 # -*- coding: utf-8 -*-
 
-import glob
-import os
 import numpy as np
 import pandas as pd
+from pathlib import Path
 from biopeaks.heart import ecg_peaks
 from wfdb.processing import compare_annotations
 
 
-GUDB_dir = r"...\experiment_data"
+data_dir = Path(".../experiment_data")    # replace with your local "experiment_data" directory once you've downloaded the database
 
-os.chdir(GUDB_dir)
-
-records = []
-annotations = []
-condition = "hand_bike"
-
-for subject in glob.glob("subject_*"):
-    records.append(os.path.join(subject, condition, "ECG.tsv"))
-    annotations.append(os.path.join(subject, condition,
-                                    "annotation_cables.tsv"))
-
+condition = "hand_bike"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
 sfreq = 250
-tolerance = 1    # tolerance must be in samples for wfdb
+tolerance = 1    # in samples
 print(f"Setting tolerance for match between algorithmic and manual annotation"
       f" to 1 sample, corresponding to {1 / sfreq} seconds at a sampling rate of {sfreq}.")
 
 sensitivity = []
 precision = []
 
-for record, annotation in zip(records, annotations):
+for subject in data_dir.glob("subject_*"):
+
+    record = subject.joinpath(f"{condition}/ECG.tsv")
+    annotation = subject.joinpath(f"{condition}/annotation_cables.tsv")
 
     try:
         ecg = np.ravel(pd.read_csv(record, sep=" ", usecols=[1],
@@ -52,14 +44,15 @@ for record, annotation in zip(records, annotations):
         fp = comparitor.fp
         fn = comparitor.fn
 
-        # plt.figure()
-        # plt.plot(ecg)
-        # plt.scatter(manupeaks, ecg[manupeaks], c="m")
-        # plt.scatter(algopeaks, ecg[algopeaks], c='g', marker='X', s=150)
-
         sensitivity.append(float(tp) / (tp + fn))
         precision.append(float(tp) / (tp + fp))
-        print(f"sensitivity = {sensitivity[-1]}, precision = {precision[-1]}")
 
-print(f"mean precision = {np.mean(precision)}, std precision = {np.std(precision)}")
-print(f"mean sensitivity = {np.mean(sensitivity)}, std sensitivity = {np.std(sensitivity)}")
+        print(f"\nResults {subject}")
+        print("-" * len(str(subject)))
+        print(f"sensitivity = {sensitivity[-1]}")
+        print(f"precision = {precision[-1]}")
+
+print(f"\nAverage results over {len(precision)} records")
+print("-" * 31)
+print(f"sensitivity: mean = {np.mean(sensitivity)}, std = {np.std(sensitivity)}")
+print(f"precision: mean = {np.mean(precision)}, std = {np.std(precision)}")
