### 2020-09-22 JanCBrammer/biopeaks#14 @JanCBrammer
> I cannot find "signal" and "annotations" as required in `benchmark_PPG.py`. where am I going wrong?

@sappelhoff, you're right. I didn't document that the "signal" and "annotations" need to be extracted from the .mat files before running `benchmark_PPG`. I updated the script such that the PPG benchmark can now be run directly on `TBME2013-PPGRR-Benchmark_R3/data` (the extraction of "signal" and "annotations" is now taken care of in the script). See 41d7a6fee470234fe8aa49774420a661f4c4beaa.

I removed the commented code from `benchmark_ECG_local` (I only used it for debugging).

Will address the module level docstrings once I get around to #11.
