+++ environment.yml
@@ -2,9 +2,9 @@ name: ci_env
 channels:
   - conda-forge
 dependencies:
-  - scipy
-  - numpy
-  - matplotlib
-  - pyside2
-  - pandas
+  - scipy>=1.4.1
+  - numpy>=1.18.1
+  - matplotlib>=3.2.1
+  - pyside2>=5.13.2
+  - pandas>=0.25.3
   - pytest-qt
