### 2020-09-23 f678f38
Updated NumPy reference.
