+++ paper/paper.bib
@@ -53,19 +53,24 @@
 
 
 @article{numpy,
-	title = {The {NumPy} Array: A Structure for Efficient Numerical Computation},
-	volume = {13},
-	issn = {1558-366X},
-	doi = {10.1109/MCSE.2011.37},
-	shorttitle = {The {NumPy} Array},
-	abstract = {In the Python world, {NumPy} arrays are the standard representation for numerical data and enable efficient implementation of numerical computations in a high-level language. As this effort shows, {NumPy} performance can be improved through three techniques: vectorizing calculations, avoiding copying data in memory, and minimizing operation counts.},
-	pages = {22--30},
-	number = {2},
-	journaltitle = {Computing in Science Engineering},
-	author = {van der Walt, Stefan and Colbert, S. Chris and Varoquaux, Gael},
-	date = {2011-03},
-	note = {Conference Name: Computing in Science Engineering},
-	keywords = {Arrays, Computational efficiency, data structures, Finite element methods, high level language, high level languages, mathematics computing, numerical analysis, Numerical analysis, numerical computation, numerical computations, numerical data, {NumPy}, numpy array, Performance evaluation, programming libraries, Python, Python programming language, Resource management, scientific programming, Vector quantization},
+	title = {Array programming with {NumPy}},
+	volume = {585},
+	copyright = {2020 The Author(s)},
+	issn = {1476-4687},
+	url = {https://www.nature.com/articles/s41586-020-2649-2},
+	doi = {10.1038/s41586-020-2649-2},
+	abstract = {Array programming provides a powerful, compact and expressive syntax for accessing, manipulating and operating on data in vectors, matrices and higher-dimensional arrays. NumPy is the primary array programming library for the Python language. It has an essential role in research analysis pipelines in fields as diverse as physics, chemistry, astronomy, geoscience, biology, psychology, materials science, engineering, finance and economics. For example, in astronomy, NumPy was an important part of the software stack used in the discovery of gravitational waves1 and in the first imaging of a black hole2. Here we review how a few fundamental array concepts lead to a simple and powerful programming paradigm for organizing, exploring and analysing scientific data. NumPy is the foundation upon which the scientific Python ecosystem is constructed. It is so pervasive that several projects, targeting audiences with specialized needs, have developed their own NumPy-like interfaces and array objects. Owing to its central position in the ecosystem, NumPy increasingly acts as an interoperability layer between such array computation libraries and, together with its application programming interface (API), provides a flexible framework to support the next decade of scientific and industrial analysis.},
+	language = {en},
+	number = {7825},
+	urldate = {2020-09-23},
+	journal = {Nature},
+	author = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del Río, Jaime Fernández and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
+	month = sep,
+	year = {2020},
+	note = {Number: 7825
+Publisher: Nature Publishing Group},
+	pages = {357--362},
[…]
