### 2020-09-23 05b446c
Made clear that Einthoven lead II is an ECG channel.
