### 2020-09-23 2b3705e
Made "Getting started" examples more concrete.
