+++ docs/user_guide.md
@@ -5,58 +5,82 @@
 ![blank](images/screenshot_blank.png)
 
 In the **menubar**, you can find the sections **_biosignal_**, **_peaks_**,
-and **_statistics_**. These contain methods for the interaction with your biosignals. On the left 
+and **_statistics_**. These contain methods for the interaction with your biosignals. On the left
 side, there's a panel containing **configurations** that allows you to customize your workflow.
 You can toggle the visibility of the **configurations** with **_show/hide configurations_** in the **menubar**.
 To the right of the **configurations** is the **datadisplay** which consists of three panels. The upper
 panel contains the biosignal as well as peaks identified in the biosignal, while the middle panel can be used
-to optionally display a marker channel. The lower panel contains any statistics 
+to optionally display a marker channel. The lower panel contains any statistics
 derived from the peaks. Adjust the height of the panels by dragging the segmenters between them up or down.
 Beneath the **datadisplay**, in the lower left corner, you find the **displaytools**. These allow you to interact with the
 biosignal. Have a look in the [functionality section](#functionality) for details on these elements.
 
 
 ## Getting started
-The following work-flow is meant as an introduction to the interface. Many other
-work flows are possible and might make more sense given your 
-requirements. Note that `biopeaks` works with the OpenSignals file format
-as well as EDF files and plain text files (.txt, .csv, .tsv) containing biosignal channels as columns. The functions
+The following work-flows are meant as an introduction to the interface. Many other
+work-flows are possible. Note that `biopeaks` works with the OpenSignals file format
+as well as EDF files, and plain text files (.txt, .csv, .tsv) containing biosignal channels as columns. The functions
 used in the exemplary work-flow are described in detail in the [functionality section](#functionality).
 
-### examplary workflow on a single file
-Before you start, select the desired options in the **configurations**. Make sure
-that **_processing options_** -> _mode_  is set to "single file" and
-[load the biosignal](#load-biosignal) to visually check its quality using the [**displaytools**](#displaytools).
-Next, you could [segment the biosignal](#segment-biosignal) based on a specific time interval
-or events in the markers. Now you can [identify the peaks](#find-peaks) in the biosignal.
-If the quality of the biosignal is sufficient, the peaks should be placed in the
-correct locations. However, if there are noisy intervals in the biosignal, peaks might be
-misplaced or not detected at all (i.e., false positives or false negatives).
+### extracting ECG features from an OpenSignals file
 
-![noise](images/screenshot_noise.png)
+1. Download the [example data](https://github.com/JanCBrammer/biopeaks/blob/master/biopeaks/docs/example_data.txt) (click the `Download` button on the upper right).
 
-If this is the case you can [edit the peak locations](#edit-peaks). Once you are
-confident that all the peaks are placed correctly you can [calculate statistics](#calculate-statistics).
-Finally, you can [save the biosignal](#save-biosignal), [peaks](#save-peaks), and/or
-[statistics](#save-statistics), depending on your requirements. If you have
-segmented the biosignal it is a good idea to save it so you can reproduce the work-flow
-later if necessary. Also, save the peaks if you're planning on [reloading](#load-peaks) them later
-or using them for further analyses (e.g., heart rate variability).
+2. Configure the processing options in the **configurations**:
+    * Since we want to analyze ECG data, make sure that **_processing options_** -> _modality_ is set to "ECG".
+    * Set **_processing options_** -> _mode_  to "single file".
[…]
