### 2020-09-23 JanCBrammer/biopeaks#12 @JanCBrammer
@sappelhoff, see 2b3705e32be27c690a930cf7128060404237258b.
