### 2020-09-25 JanCBrammer/biopeaks#9 @TomDonoghue
I think these are great suggestions, and on the topic of adding direct links of brief summaries to the README, I might add a note on installation. Some brief info on installation is something I typically expect in the README.

If I'm on Github, I basically just want to know "is this on PYPI". One option is to add a PYPI badge-link to the README (an example of this, and other badges here: https://github.com/neurodsp-tools/neurodsp).

You might also just add a direct link from README -> installation page, and/or a mini section on the README that includes info, even if that's just something like
```
`biopeaks` can be installed from pypi as
`pip install biopeaks`

For more information on installation, see the documentation page [link]
```

Note: I'm also from the JOSS review :)
