### 2020-09-25 JanCBrammer/biopeaks#11 @TomDonoghue
I totally agree on this point - I think picking and applying a consistent documentation structure would help the codebase a fair amount. It is true that the average user might not see this (since it's a GUI), but under the assumption that this is a tool to be supported for a while, the documentation is for maintainers (including you in the future). I reckon it will be much easier to learn whats going on with nice in-code documentation, and this would help a lot with onboarding (right now, just scrolling through the code, it's harder to get a sense of how things work).

Note: I'm also from the JOSS review
