### 2020-09-25 7c1d10b
Added sections on installation, documentation, and contributing.
