+++ README.md
@@ -1,12 +1,8 @@
 ![logo](docs/images/logo.png)
 
-![](https://github.com/JanCBrammer/biopeaks/workflows/test/badge.svg?branch=dev)
-
-
-# Citation
-Click on the badge below to cite `biopeaks` in a format of your choice.
-
+![GH Actions](https://github.com/JanCBrammer/biopeaks/workflows/test/badge.svg?branch=dev)
 [![DOI](https://www.zenodo.org/badge/172897525.svg)](https://www.zenodo.org/badge/latestdoi/172897525)
+[![PyPI version](https://img.shields.io/pypi/v/biopeaks.svg)](https://pypi.org/project/biopeaks/)
 
 
 # General Information
@@ -27,6 +23,37 @@ peaks in breathing signals) with signal-specific, sensible defaults
 + .csv export of extrema and instantaneous features for further analysis (e.g., heart rate variability)
 + automatic analysis of multiple files (batch processing)
 
-Visit the [documentation](https://jancbrammer.github.io/biopeaks/) for additional information.
 
 ![GUI](docs/images/screenshot_statistics.png)
+
+
+# Installation
+
+`biopeaks` can be installed from PyPI:
+
+```
+pip install biopeaks
+```
+You can find more details on the installation [here](https://jancbrammer.github.io/biopeaks/installation.html).
+
+
+# Documentation
+
+Have a look at the [user guide](https://jancbrammer.github.io/biopeaks/user_guide.html) to get started with `biopeaks`.
+
+
+# Contributors welcome!
+
+Improvements or additions to the repository (documentation, tests, code) are welcome and encouraged.
+Spotted a typo in the documentation? Caught a bug in the code? Ideas for improving the documentation,
+increase test coverage, or adding features to the GUI? Get started with the [contributor guide](https://jancbrammer.github.io/biopeaks/contributor_guide.html).
+
+
+# Citation
+
+Click [here](https://zenodo.org/record/3971650#.X24GX-3gphE) to cite `biopeaks` in a format of your choice.
+Look for the `Cite as` box on the lower right of the page.
+
+
+
+
