### 2020-09-25 c2092bb
Made contributor guide more explicit.
