### 2020-09-25 JanCBrammer/biopeaks#9 @JanCBrammer
@sappelhoff, @TomDonoghue, thanks for your advise and suggestions.

The README now has
* a more prominent link to the documentation
* a short section on installation and corresponding link to more detailled instructions
* a link to the contributor guide

See 7c1d10b25ab109ae6f8466d55c5009a789d23f67.

The contributor guide has been improved with more elaborate instructions (c2092bbef1dcc97515a73337fe8dac4243c24f8b).

I did some research on the code of conduct and added one to the repository (884bf6a2320839c9ac4db92a66244418d63fe345). The contributor/community aspect was not on my radar so much until now, since the development was mostly a solitary effort so far. But I hope for this to change in the future. Hence I appreciate you bringing the contributor/community aspect to my attention.
