### 2020-09-27 JanCBrammer/biopeaks#9 @sappelhoff
Excellent, thanks. @TomDonoghue please feel free to re-open in case there were points that you feel have not been addressed.
