### 2020-09-28 JanCBrammer/biopeaks#12 @sappelhoff
Cool, I just tried it out and I am very happy with the improvements, great job @JanCBrammer!
