### 2020-09-28 JanCBrammer/biopeaks#15 @sappelhoff
custom zoom gets undone during manual peak editing + other issues
related to JOSS review (https://github.com/openjournals/joss-reviews/issues/2621)

some notes while trying the new getting started section from https://github.com/JanCBrammer/biopeaks/issues/12:

- after editing some peaks and specifically while adding a new one manually (click and press "a"), my currently "zoomed in" window is re-sized to the full window again and I have to zoom back in ... is this an error you also observe? Or is it linux (ubuntu 18.04) specific?
- manually edited peaks are indistinguishable from the automatically selected peaks --> both show up as purple dots for me. It'd be nice to be able to distinguish them
- manually edited peaks that I put into the "wrong" place are not "autocorrected" when pressing the peak autocorrection button ... is that intended?
