### 2020-09-28 JanCBrammer/biopeaks#14 @sappelhoff
great, I was able to run all benchmarks now. Thanks a bunch.

The performance details you describe in the paper match approximately with what I can reproduce on my machine (slight rounding deviations are to be expected I guess.)
