### 2020-09-28 #2621 @sappelhoff
As an update: @JanCBrammer has addressed most of my suggestions by now. There are three remaining open issues from my side:

- consistent docstring documentation
- test coverage reporting
- developer documentation on how to build the docs

It'd be nice to have these points addressed before a final evaluation.
