### 2020-09-29 JanCBrammer/biopeaks#16 @TomDonoghue
Small suggested paper edits
As part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/2621), I've been reading the paper, with some small notes.

Quick note:
- it seems a bit surprising to me that mean precision and sensitivity values are exactly the same (in the table). This could of course be correct, but it might be worth double checking there wasn't a copy/paste error here (?)

Small notes for the references:
- I don't think the Elgendi paper needs to reference the editor (or, at least, it is inconsistent for most papers you don't do this). The McKinney paper also lists editors, but from the PDF I'm not the listed editors are accurate, so I might check and remove there as well.
- the 'Howell' paper is the only one that indicates a month as part of the date, and I think it probably doesn't need to do so
- I believe the numpy citation should have the first author's last name as `van der Walt` (all part of the last name). It seems to have been parsed to put the 'van der' as a middle name
