### 2020-09-29 JanCBrammer/biopeaks#8 @TomDonoghue
Somewhat related: it might be worth adding a note somewhere on the requirements for running the tests locally. As well as the standard of needing `pytest`, when trying to run the tests I got an error that `fixture 'qtbot' not found`, which I tracked to being because `pytest-qt` also being required. This isn't clearly noted anywhere.

Somewhere, I would add a note of test-requirements, and maybe some instructions on running tests locally (this is perhaps most relevant in a contributor guide, as you would likely want contributors to run and check tests.

As a note, I was able to run tests locally. I'm not sure of the coverage (adding that is a good idea), though the GUI tests looked (in terms of how many things popped up) fairly comprehensive.
