### 2020-09-29 #2621 @TomDonoghue
Just a note that I have basically finished my first round of review. Thanks to @sappelhoff for making my job easier, as he already pointed out a lot of good suggestions! I made some additional suggestions, mostly by opening a couple issues on the tool repo. As a general note, I consider a few of the issues I opened as _suggestions_ (not strictly required for the paper), and so to guide what I think is most important would be:
- [ ] I agree with @sappelhoff in particular on fixing up the documentation a bit (https://github.com/JanCBrammer/biopeaks/issues/11)
- [ ] some small paper checks & fixes (https://github.com/JanCBrammer/biopeaks/issues/16)
- [ ] noting test dependencies and how to run them, which I've added to an existing issue (https://github.com/JanCBrammer/biopeaks/issues/8)

The other issues are potential clean ups / clarifications / extensions that may be useful for the module, but are perhaps a little longer term, and I don't think necessarily need to be done now.

When these things are all checked / done, I'm happy to come back and sign-off on the remaining points.
