### 2020-09-29 #2621 @JanCBrammer
@sappelhoff, @TomDonoghue, thank you for you clear and thorough reviews! I'll try to address your issues over the coming two weeks. I'll keep you in the loop by tagging you once I've resolved an issue. Also, I will notify you here once I think the repo is ready for your final evaluation / sign-off (in case you don't have time and/or resources to keep track of single issues).
