### 2020-09-29 6685c6e
Included instructions for running tests and benchmarks in contributor guide.
