### 2020-09-29 JanCBrammer/biopeaks#8 @JanCBrammer
> it might be worth adding a note somewhere on the requirements for running the tests locally

@TomDonoghue, I've included [testing instructions](https://jancbrammer.github.io/biopeaks/contributor_guide.html#tests) in the contributor guide (see 6685c6e6f9a49e8de59ffb2fd92fe4be1f30ca46).
