### 2020-09-29 cb89f3a
Made benchmarking scripts more consistent and rounded average results to four decimal places.
