+++ biopeaks/benchmarks/benchmark_ECG_local.py
@@ -9,7 +9,7 @@ from wfdb.processing import compare_annotations
 
 data_dir = Path(".../experiment_data")    # replace with your local "experiment_data" directory once you've downloaded the database
 
-condition = "hand_bike"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
+condition = "sitting"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
 sfreq = 250
 tolerance = 1    # in samples
 print(f"Setting tolerance for match between algorithmic and manual annotation"
@@ -44,8 +44,8 @@ for subject in data_dir.glob("subject_*"):
         fp = comparitor.fp
         fn = comparitor.fn
 
-        sensitivity.append(float(tp) / (tp + fn))
-        precision.append(float(tp) / (tp + fp))
+        sensitivity.append(tp / (tp + fn))
+        precision.append(tp / (tp + fp))
 
         print(f"\nResults {subject}")
         print("-" * len(str(subject)))
@@ -54,5 +54,5 @@ for subject in data_dir.glob("subject_*"):
 
 print(f"\nAverage results over {len(precision)} records")
 print("-" * 31)
-print(f"sensitivity: mean = {np.mean(sensitivity)}, std = {np.std(sensitivity)}")
-print(f"precision: mean = {np.mean(precision)}, std = {np.std(precision)}")
+print(f"sensitivity: mean = {round(np.mean(sensitivity), 4)}, std = {round(np.std(sensitivity), 4)}")
+print(f"precision: mean = {round(np.mean(precision), 4)}, std = {round(np.std(precision), 4)}")
