### 2020-09-29 416ad91
Removed editor field from elgendi and pandas references and month field from gudb reference.
