+++ paper/paper.bib
@@ -79,7 +79,6 @@ Publisher: Nature Publishing Group},
 	pages = {56 -- 61},
 	booktitle = {Proceedings of the 9th Python in Science Conference},
 	author = {{McKinney}, Wes},
-	editor = {van der Walt, Stéfan and Millman, Jarrod},
 	date = {2010},
 	doi = {10.25080/Majora-92bf1922-00a}
 }
@@ -199,7 +198,6 @@ Conclusions: The {HRV} artefact correction algorithm presented in this study pro
 	urldate = {2020-08-14},
 	journal = {PLoS ONE},
 	author = {Elgendi, Mohamed and Norton, Ian and Brearley, Matt and Abbott, Derek and Schuurmans, Dale},
-	editor = {Bondarenko, Vladimir E.},
 	month = oct,
 	year = {2013},
 	pages = {e76585},
@@ -245,7 +243,6 @@ The cheststrap ECG allowed R peak detection even while jogging at a very high pr
 	language = {en},
 	urldate = {2020-08-14},
 	author = {Howell, Luis and Porr, Bernd},
-	month = dec,
 	year = {2018},
 	note = {Publisher: University of Glasgow},
 }
