### 2020-09-29 JanCBrammer/biopeaks#16 @JanCBrammer
>     * it seems a bit surprising to me that mean precision and sensitivity values are exactly the same (in the table)

@TomDonoghue, I now round the precision and sensitivity values to four instead of three decimal places, which makes it clear that they're different (cb89f3aa9338fdebc3f264b9cb04cb87ffa945b5).

I also formatted the references according to your suggestions: 416ad91766ca8e07ed3437f1ecc1f807e8e2fb03. You should be able to have a look at the updated paper preview [here](http://res.cloudinary.com/hju22ue2k/image/upload/v1601393483/g4nmuuw4lqkf91zjuihy.pdf) (let me know if the link is broken). I also replaced the old NumPy reference with the (now somewhat infamous) 2020 NumPy paper.
