+++ environment.yml
@@ -8,3 +8,4 @@ dependencies:
   - pyside2>=5.13.2
   - pandas>=0.25.3
   - pytest-qt
+  - pytest-cov
