+++ .github/workflows/test.yml
@@ -137,11 +137,16 @@ jobs:
     - name: Test with pytest on Ubuntu
       if: matrix.os == 'ubuntu-latest'
       shell: bash -l {0}
-      run: |
+      run: | # generate coverage report on ubuntu only
         sudo apt-get install -y xvfb x11-utils libxkbcommon-x11-0
-        xvfb-run pytest -v
+        xvfb-run pytest -v --cov-config=.coveragerc --cov-report=xml --cov=biopeaks/
     - name: Test with pytest on Windows and MacOS
       if: matrix.os == 'windows-latest' || matrix.os == 'macos-latest'
       shell: bash -l {0}
       run: |
         pytest -v
+    - name: Upload coverage report
+      if: matrix.os == 'ubuntu-latest'
+      uses: codecov/codecov-action@v1
+      with:
+        file: ./coverage.xml
