### 2020-09-30 5cf56ac
Setting up codecov (1): Only upload report from Python 3.7 run.
