### 2020-10-01 a5e8022
Test source and wheel only before publishing to PyPI (1).
