### 2020-10-01 fc2e9e2
Test source and wheel only before publishing to PyPI (2).
