### 2020-10-01 5d1e3ea
Test source and wheel only before publishing to PyPI (3).
