### 2020-10-01 JanCBrammer/biopeaks#8 @JanCBrammer
> It would be great to add a tool that measures what proportion of your code is actually covered by the tests.

@sappelhoff, I now collect and report code coverage. However, the coverage is deceptively low (59%) since pytest-cov does not seem to pick up on code that is run within threads (see [here](https://codecov.io/gh/JanCBrammer/biopeaks/src/master/biopeaks/controller.py)). Looking into it.
