+++ README.md
@@ -56,5 +56,10 @@ Click [here](https://zenodo.org/record/3971650#.X24GX-3gphE) to cite `biopeaks` 
 Look for the `Cite as` box on the lower right of the page.
 
 
+# Changelog
+
+Have a look at the [changelog](https://jancbrammer.github.io/biopeaks/changelog.html) to get an overview of what has changed throughout the versions of `biopeaks`.
+
+
 
 
