### 2020-10-05 f2d6101
Added section about documentation.
