### 2020-10-05 JanCBrammer/biopeaks#10 @JanCBrammer
@sappelhoff, see f2d61019eb9c59ebc253707f38d3dfc025276aa0.
