### 2020-10-05 JanCBrammer/biopeaks#10 @sappelhoff
looks good to me, thanks!
