### 2020-10-05 f549e72
Made compute_threshold and update_indices private functions in heart module since they're not called anywhere else.
