### 2020-10-05 e420311
Worked on docstrings.
