### 2020-10-06 f5acffe
Format anchor tags.
