+++ docs/contributor_guide.md
@@ -26,7 +26,7 @@ Follow these steps to contribute documentation, tests, or code:
    1. Make a [local clone of your fork](https://docs.github.com/en/free-pro-team@latest/github/getting-started-with-github/fork-a-repo#step-2-create-a-local-clone-of-your-fork).
    2. Navigate to the directory containing the cloned fork.
    3. Install `biopeaks` with `pip install -e .` The `e` stand for editable, meaning you can immediatly test all the changes you make to the cloned fork. The `.` simply tells pip to install the content of the current directory.
-4. Implement your contribution in the `topic` branch, following the [conventions](##Conventions).
+4. Implement your contribution in the `topic` branch, following the [conventions](#conventions).
 5. [Make a pull request](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request-from-a-fork) from the `topic` branch on your fork to the [`dev` branch of the `biopeaks` repository](https://github.com/JanCBrammer/biopeaks/tree/dev).
 6. Once all tests pass and your contribution has been reviewed, your PR will merged and you'll be added to the list of contributors.
 
@@ -38,7 +38,7 @@ Follow these steps to contribute documentation, tests, or code:
 * avoid introducing new dependencies
 * write [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html) docstrings
   for every (non-private) new function
-* add tests if you contribute code that is not covered by the existing [tests](##Tests)
+* add tests if you contribute code that is not covered by the existing [tests](#tests)
 
 ### Code style
 
