### 2020-10-06 070d76b
Removed unused period attribute.
