+++ biopeaks/model.py
@@ -1,4 +1,9 @@
 # -*- coding: utf-8 -*-
+"""Model component of the MVC implementation.
+
+Stores the current state of the application. Receives state updates from the
+View or Controller and informs View about changes in state.
+"""
 
 import numpy as np
 from pathlib import Path
@@ -7,7 +12,6 @@ from PySide2.QtCore import QObject, Signal, Slot, Property
 
 class Model(QObject):
 
-    # costum signals
     signal_changed = Signal(object)
     peaks_changed = Signal(object)
     marker_changed = Signal(object)
@@ -21,15 +25,15 @@ class Model(QObject):
     progress_changed = Signal(int)
     model_reset = Signal()
 
-    # the following model attributes aren't slots, but are set by controller
-    # methods
+    # The following attributes are set by the View or Controller (i.e., they
+    # are not slots connected to a signal).
 
     @property
     def savestats(self):
         return self._savestats
 
     @savestats.setter
-    def savestats(self, value):
+    def savestats(self, value):    # View
         self._savestats = value
 
     @property
@@ -37,11 +41,11 @@ class Model(QObject):
         return self._signal
 
     @signal.setter
-    def signal(self, value):
+    def signal(self, value):    # Controller
         self._signal = value
-        # check if attribute is reset to None, in that case do not emit its
-        # signal, also do not emit signal if plotting is not desired (e.g.,
-        # during batch processing)
+        # Check if attribute is reset to None, in that case do not emit its
+        # signal. Also do not emit signal if plotting is not desired (e.g.,
+        # during batch processing).
         if value is not None and self.plotting:
             self.signal_changed.emit(value)
 
@@ -50,7 +54,7 @@ class Model(QObject):
         return self._peaks
 
     @peaks.setter
-    def peaks(self, value):
+    def peaks(self, value):    # Controller
 
         if isinstance(value, np.ndarray) and value.size > 1:
             self._peaks = value
@@ -58,19 +62,11 @@ class Model(QObject):
             self.peaks_changed.emit(value)
 
     @property
-    def period(self):
-        return self._period
-
-    @period.setter
-    def period(self, value):
-        self._period = value
-
-    @property
     def periodintp(self):
         return self._periodintp
 
     @periodintp.setter
-    def periodintp(self, value):
+    def periodintp(self, value):    # Controller
         self._periodintp = value
         if value is not None and self.plotting:
             self.period_changed.emit(value)
@@ -80,7 +76,7 @@ class Model(QObject):
         return self._rateintp
 
     @rateintp.setter
-    def rateintp(self, value):
+    def rateintp(self, value):    # Controller
         self._rateintp = value
         if value is not None and self.plotting:
             self.rate_changed.emit(value)
@@ -90,7 +86,7 @@ class Model(QObject):
         return self._tidalampintp
 
     @tidalampintp.setter
-    def tidalampintp(self, value):
+    def tidalampintp(self, value):    # Controller
         self._tidalampintp = value
         if value is not None and self.plotting:
             self.tidalamp_changed.emit(value)
@@ -100,7 +96,7 @@ class Model(QObject):
         return self._sec
 
     @sec.setter
-    def sec(self, value):
+    def sec(self, value):    # Controller
         self._sec = value
 
     @property
@@ -108,7 +104,7 @@ class Model(QObject):
         return self._marker
 
     @marker.setter
-    def marker(self, value):
+    def marker(self, value):    # Controller
         self._marker = value
         if value is not None and self.plotting:
             # In case the marker channel is sampled at a different rate than
@@ -126,7 +122,7 @@ class Model(QObject):
         return self._rpathsignal
 
     @rpathsignal.setter
[…]
