### 2020-10-06 525bb43
Implement all model attributes with getter and setter.
