### 2020-10-06 dc06e13
Worked on docstrings.
