+++ biopeaks/model.py
@@ -95,50 +95,68 @@ class Model(QObject):
 
     @property
     def plotting(self):
+        """
+        Set by Controller.
+        """
         return self._plotting
 
     @plotting.setter
-    def plotting(self, value):    # Controller
+    def plotting(self, value):
         self._plotting = value
 
     @property
     def loaded(self):
+        """
+        Set by Controller.
+        """
         return self._loaded
 
     @loaded.setter
-    def loaded(self, value):    # Controller
+    def loaded(self, value):
         self._loaded = value
 
     @property
     def sfreqmarker(self):
+        """
+        Set by Controller.
+        """
         return self._sfreqmarker
 
     @sfreqmarker.setter
-    def sfreqmarker(self, value):    # Controller
+    def sfreqmarker(self, value):
         self._sfreqmarker = value
 
     @property
     def sfreq(self):
+        """
+        Set by Controller.
+        """
         return self._sfreq
 
     @sfreq.setter
-    def sfreq(self, value):    # Controller
+    def sfreq(self, value):
         self._sfreq = value
 
     @property
     def savestats(self):
+        """
+        Set by View.
+        """
         return self._savestats
 
     @savestats.setter
-    def savestats(self, value):    # View
+    def savestats(self, value):
         self._savestats = value
 
     @property
     def signal(self):
+        """
+        Set by Controller.
+        """
         return self._signal
 
     @signal.setter
-    def signal(self, value):    # Controller
+    def signal(self, value):
         self._signal = value
         # Check if attribute is reset to None, in that case do not emit its
         # signal. Also do not emit signal if plotting is not desired (e.g.,
@@ -148,10 +166,13 @@ class Model(QObject):
 
     @property
     def peaks(self):
+        """
+        Set by Controller.
+        """
         return self._peaks
 
     @peaks.setter
-    def peaks(self, value):    # Controller
+    def peaks(self, value):
 
         if isinstance(value, np.ndarray) and value.size > 1:
             self._peaks = value
@@ -160,48 +181,63 @@ class Model(QObject):
 
     @property
     def periodintp(self):
+        """
+        Set by Controller.
+        """
         return self._periodintp
 
     @periodintp.setter
-    def periodintp(self, value):    # Controller
+    def periodintp(self, value):
         self._periodintp = value
         if value is not None and self._plotting:
             self.period_changed.emit(value)
 
     @property
     def rateintp(self):
+        """
+        Set by Controller.
+        """
         return self._rateintp
 
     @rateintp.setter
-    def rateintp(self, value):    # Controller
+    def rateintp(self, value):
         self._rateintp = value
         if value is not None and self._plotting:
             self.rate_changed.emit(value)
 
     @property
     def tidalampintp(self):
+        """
+        Set by Controller.
+        """
         return self._tidalampintp
 
     @tidalampintp.setter
-    def tidalampintp(self, value):    # Controller
+    def tidalampintp(self, value):
         self._tidalampintp = value
         if value is not None and self._plotting:
             self.tidalamp_changed.emit(value)
 
     @property
     def sec(self):
+        """
+        Set by Controller.
+        """
         return self._sec
 
     @sec.setter
-    def sec(self, value):    # Controller
+    def sec(self, value):
         self._sec = value
 
     @property
     def marker(self):
+        """
+        Set by Controller.
+        """
         return self._marker
 
     @marker.setter
-    def marker(self, value):    # Controller
+    def marker(self, value):
         self._marker = value
         if value is not None and self._plotting:
             # In case the marker channel is sampled at a different rate than
@@ -216,10 +252,13 @@ class Model(QObject):
 
     @property
[…]
