### 2020-10-08 66fda94
Made dispatcher private.
