### 2020-10-08 84509c6
Worked on docstrings.
