+++ biopeaks/model.py
@@ -1,9 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Model component of the MVC implementation.
-
-Stores the current state of the application. Receives state updates from the
-View or Controller and informs View about changes in state.
-"""
+"""Model component of the MVC application."""
 
 import numpy as np
 from pathlib import Path
@@ -11,12 +7,75 @@ from PySide2.QtCore import QObject, Signal, Slot, Property
 
 
 class Model(QObject):
+    """Model component of the MVC application.
+
+    Stores the current state of the application. Receives state updates from
+    the View or Controller and informs View about changes in state.
+
+    Attributes
+    ----------
+    signal
+    peaks
+    periodintp
+    rateintp
+    tidalampintp
+    sec
+    marker
+    segment
+    status
+    progress
+    sfreq
+    sfreqmarker
+    loaded
+    plotting
+    signalchan
+    markerchan
+    modality
+    batchmode
+    peakseditable
+    fpaths
+    wpathpeaks
+    wdirpeaks
+    rpathpeaks
+    wpathsignal
+    rpathsignal
+    wpathstats
+    wdirstats
+    savebatchpeaks
+    correctbatchpeaks
+    savestats
+    filetype
+    customheader
+    signal_changed : Signal
+        Notify View that the signal attribute changed.
+    peaks_changed : Signal
+        Notify View that the peaks attribute changed.
+    marker_changed : Signal
+        Notify View that the marker attribute changed.
+    segment_changed : Signal
+        Notify View that the segment attribute changed.
+    period_changed : Signal
+        Notify View that the period attribute changed.
+    rate_changed : Signal
+        Notify View that the rate attribute changed.
+    tidalamp_changed : Signal
+        Notify View that the tidalamp attribute changed.
+    path_changed : Signal
+        Notify View that the rpathsignal attribute changed.
+    status_changed : Signal
+        Notify View that the status attribute changed.
+    progress_changed : Signal
+        Notify View that the progress attribute changed.
+    model_reset : Signal
+        Notify View that attributes have been reset to their default value.
+        See reset method.
+
+    """
 
     signal_changed = Signal(object)
     peaks_changed = Signal(object)
     marker_changed = Signal(object)
     segment_changed = Signal(object)
-    # makes sure that the emitted stats signal has the same length as self._sec
     period_changed = Signal(object)
     rate_changed = Signal(object)
     tidalamp_changed = Signal(object)
@@ -26,8 +85,14 @@ class Model(QObject):
     model_reset = Signal()
 
     def __init__(self):
+        """Instantiate all attributes with their default value."""
         super().__init__()
 
+        self._loaded = False
+        self._peakseditable = False
+        self._plotting = True
+        self._savebatchpeaks = False
+        self._correctbatchpeaks = False
         self._signal = None
         self._peaks = None
         self._periodintp = None
@@ -40,13 +105,10 @@ class Model(QObject):
         self._progress = None
         self._sfreq = None
         self._sfreqmarker = None
-        self._loaded = False
-        self._plotting = True
         self._signalchan = None
         self._markerchan = None
         self._modality = None
         self._batchmode = None
-        self._peakseditable = False
         self._fpaths = None
         self._wpathpeaks = None
         self._wdirpeaks = None
@@ -55,20 +117,24 @@ class Model(QObject):
         self._rpathsignal = None
         self._wpathstats = None
         self._wdirstats = None
-        self._savebatchpeaks = False
-        self._correctbatchpeaks = False
-        self._savestats = {"period": False, "rate": False, "tidalamp": False}
         self._filetype = None
+        self._savestats = {"period": False, "rate": False, "tidalamp": False}
         self._customheader = {"signalidx": None, "markeridx": None,
[…]
