### 2020-10-09 fbe77d4
Worked on docstrings.
