+++ biopeaks/view.py
@@ -17,6 +17,17 @@ import biopeaks.resources    # noqa
 
 
 class CustomNavigationToolbar(NavigationToolbar):
+    """Matplotlib navigation toolbar.
+
+    Same as superclass except for removed Subplot-configuration and Save
+    buttons.
+
+    See Also
+    --------
+    matplotlib.backends.backend_qt5agg.NavigationToolbar2QT
+
+    """
+
     toolitems = [t for t in NavigationToolbar.toolitems if t[0] in
                  ("Home", "Pan", "Zoom", "Back", "Forward")]    # only retain desired functionality
 
@@ -27,9 +38,11 @@ class View(QMainWindow):
     Presents the state of the application as well as the available means of
     interaction. Receives updates about the state from the Model and informs
     Controller about user interactions.
+
     """
 
     def __init__(self, model, controller):
+        """Define GUI elements and their layout."""
         super().__init__()
 
         self._model = model
@@ -37,21 +50,16 @@ class View(QMainWindow):
         self.segmentcursor = False
         self.togglecolors = {"#1f77b4": "m", "m": "#1f77b4"}
 
-
-        #################################################################
-        # define GUI layout and connect input widgets to external slots #
-        #################################################################
-
         self.setWindowTitle("biopeaks")
         self.setGeometry(50, 50, 1750, 750)
         self.setWindowIcon(QIcon(":/python_icon.png"))
 
-        # figure0 for signal
+        # Figure for biosignal.
         self.figure0 = Figure()
         self.canvas0 = FigureCanvas(self.figure0)
         # Enforce minimum height, otherwise resizing with self.splitter causes
         # mpl to throw an error because figure is resized to height 0. The
-        # widget can still be fully collapsed with self.splitter-
+        # widget can still be fully collapsed with self.splitter.
         self.canvas0.setMinimumHeight(1)    # in pixels
         self.ax00 = self.figure0.add_subplot(1, 1, 1)
         self.ax00.set_frame_on(False)
@@ -60,8 +68,7 @@ class View(QMainWindow):
         self.scat = None
         self.segmentspan = None
 
-
-        # figure1 for marker
+        # Figure for marker.
         self.figure1 = Figure()
         self.canvas1 = FigureCanvas(self.figure1)
         self.canvas1.setMinimumHeight(1)
@@ -71,8 +78,7 @@ class View(QMainWindow):
         self.figure1.subplots_adjust(left=0.04, right=0.98)
         self.line10 = None
 
-
-        # figure2 for statistics
+        # Figure for statistics.
         self.figure2 = Figure()
         self.canvas2 = FigureCanvas(self.figure2)
         self.canvas2.setMinimumHeight(1)
@@ -90,23 +96,22 @@ class View(QMainWindow):
         self.line22 = None
         self.figure2.subplots_adjust(left=0.04, right=0.98)
 
-        # navigation bar
         self.navitools = CustomNavigationToolbar(self.canvas0, self)
 
-        # peak editing
+        # Peak editing.
         self.editcheckbox = QCheckBox("editable", self)
         self.editcheckbox.stateChanged.connect(self._model.set_peakseditable)
 
-        # peak saving batch
+        # Peak saving during batch processing.
         self.savecheckbox = QCheckBox("save during batch processing", self)
         self.savecheckbox.stateChanged.connect(self._model.set_savebatchpeaks)
 
-        # peak auto-correction batch
+        # Peak auto-correction during batch processing.
         self.correctcheckbox = QCheckBox("correct during batch processing",
                                          self)
         self.correctcheckbox.stateChanged.connect(self._model.set_correctbatchpeaks)
 
-        # selecting stats for saving
+        # Selection of stats for saving.
[…]
