### 2020-10-11 c851543
Worked on docstrings.
