+++ biopeaks/controller.py
@@ -30,33 +30,64 @@ writefuncs = {"Custom": write_custom,
               "EDF": write_edf}
 
 
-# threading is implemented according to https://pythonguis.com/courses/
-# multithreading-pyqt-applications-qthreadpool/complete-example/
-class WorkerSignals(QObject):
+class WorkerSignal(QObject):
+    """Signal to be used in QRunnable.
+
+    Custom signals can only be defined on objects derived from QObject. Since
+    QRunnable is not derived from QObject, QRunnable cannot be extended with
+    signals directly.
+
+    Attributes
+    ----------
+    progress : Signal
+        Progress signal to be emitted by QRunnable.
+
+    """
 
     progress = Signal(int)
 
 
 class Worker(QRunnable):
+    """Execute a Controller method."""
 
-    def __init__(self, fn, controller, **kwargs):
+    def __init__(self, method, controller, **kwargs):
+        """Initiate with Controller instance, and -method, as well as signal.
+
+        Parameters
+        ----------
+        method : method
+            The Controller method to be executed.
+        controller : QObject
+            The Controller instance.
+        """
         super(Worker, self).__init__()
-        self.fn = fn
+        self.method = method
         self.kwargs = kwargs
-        self.signals = WorkerSignals()
+        self.signals = WorkerSignal()
         self.controller = controller
 
     def run(self):
+        """Execute method."""
         self.signals.progress.emit(0)
-        self.fn(self.controller, **self.kwargs)
+        self.method(self.controller, **self.kwargs)
         self.signals.progress.emit(1)
 
 
-# decorator that runs Controller methods in Worker thread
-def threaded(fn):
-    @wraps(fn)
+def threaded(method):
+    """Execute a Controller method in a thread.
+
+    This decorator executes a Controller method in a separate thread that is
+    managed by QThreadPool. Note that the Controller instance (i.e., self) is
+    automatically passed in as first argument to threader.
+
+    Parameters
+    ----------
+    method : method
+        A Controller method.
+    """
+    @wraps(method)
     def threader(controller, **kwargs):
-        worker = Worker(fn, controller, **kwargs)
+        worker = Worker(method, controller, **kwargs)
         worker.signals.progress.connect(controller._model.progress)
         controller.threadpool.start(worker)
     return threader
@@ -66,68 +97,83 @@ class Controller(QObject):
     """Controller component of the MVC application.
 
     Manipulates the state of the Model based on user input from the View.
+
     """
 
     def __init__(self, model):
+        """Initiate with Model and threadpool.
+
+        Parameters
+        ----------
+        model : QObject
+            Model component of the MVC application.
+        """
         super().__init__()
 
         self._model = model
         self.threadpool = QThreadPool()
         self.threadpool.setMaxThreadCount(1)
 
+    def load_channels(self):
+        """Load channels from file.
 
-    def get_fpaths(self):
-
+        Open a file dialog and let user load one or multiple files containing
+        a biosignal and/or marker channel. Depending on the configuration,
+        either load single dataset or initiate batch processing with multiple
+        datasets.
+        """
         self._model.fpaths = getOpenFileNames(None, 'Choose your data',
                                               "\\home")[0]
         if not self._model.fpaths:
-            self._model.customheader = dict.fromkeys(self._model.customheader, None)    # for custom files also reset header
             self._model.set_filetype(None)    # reset file type
+            self._model.customheader = dict.fromkeys(self._model.customheader, None)    # for custom files also reset header
             return
         if (self._model.batchmode == 'multiple files' and
             len(self._model.fpaths) >= 1):
-            self.batch_processor()
+            self._batch_processor()
         elif (self._model.batchmode == 'single file' and
               len(self._model.fpaths) == 1):
             self._model.reset()
-            self.read_channels()
+            self._load_channels()
 
[…]
