### 2020-10-12 ed5252f
Worked on docstrings.
