+++ biopeaks/filters.py
@@ -1,65 +1,199 @@
 # -*- coding: utf-8 -*-
+"""Collection of filters.
+
+To obtain a zero-phase, non-causal filter, scipy.signal.filtfilt is used to
+filter the signal. I.e. the filtered signal is not phase shifted with respect
+to the original signal since the filtering is performed in both directions
+(phase shifts cancel each other out).
+"""
 
 from scipy.signal import butter, filtfilt
 import numpy as np
 
-# use filtfilt to obtain a zero-phase filter, i.e. the filtered signal is not
-# phase shifted with respect to the original signal since the filtering is
-# performed in both directions (phase shifts cancel each other out)
 
+def butter_lowpass(cutoff, sfreq, order=5):
+    """Design IIR low-pass Butterworth filter.
 
-def butter_lowpass(cutoff, fs, order=5):
-    nyq = 0.5 * fs
+    Parameters
+    ----------
+    cutoff : float
+        Cutoff frequency. Passband is below `cutoff`.
+    sfreq : int
+        Sampling frequency of signal.
+    order : int, optional
+        Filter order, by default 5.
+
+    Returns
+    -------
+    b, a : ndarray, ndarray
+        Numerator (`b`) and denominator (`a`) polynomials of the IIR filter.
+    """
+    nyq = 0.5 * sfreq
     normal_cutoff = cutoff / nyq
     b, a = butter(order, normal_cutoff, btype='low')
     return b, a
 
 
-def butter_lowpass_filter(data, cutoff, fs, order=5):
-    b, a = butter_lowpass(cutoff, fs, order=order)
-    y = filtfilt(b, a, data, method='pad')
+def butter_lowpass_filter(signal, cutoff, sfreq, order=5):
+    """Apply IIR low-pass Butterworth filter.
+
+    Parameters
+    ----------
+    signal : ndarray
+        The signal to be filtered.
+    cutoff : float
+        Cutoff frequency. Passband is below `cutoff`.
+    sfreq : int
+        Sampling frequency of `signal`.
+    order : int, optional
+        Filter order, by default 5.
+
+    Returns
+    -------
+    y : ndarray
+        The filtered signal.
+    """
+    b, a = butter_lowpass(cutoff, sfreq, order=order)
+    y = filtfilt(b, a, signal, method='pad')
     return y
 
 
-def butter_highpass(cutoff, fs, order=5):
-    nyq = 0.5 * fs
+def butter_highpass(cutoff, sfreq, order=5):
+    """Design IIR high-pass Butterworth filter.
+
+    Parameters
+    ----------
+    cutoff : float
+        Cutoff frequency. Passband is above `cutoff`.
+    sfreq : int
+        Sampling frequency of `signal`.
+    order : int, optional
+        Filter order, by default 5.
+
+    Returns
+    -------
+    b, a : ndarray, ndarray
+        Numerator (`b`) and denominator (`a`) polynomials of the IIR filter.
+    """
+    nyq = 0.5 * sfreq
     normal_cutoff = cutoff / nyq
     b, a = butter(order, normal_cutoff, btype='high')
     return b, a
 
 
-def butter_highpass_filter(data, cutoff, fs, order=5):
-    b, a = butter_highpass(cutoff, fs, order=order)
-    y = filtfilt(b, a, data, method='pad')
+def butter_highpass_filter(signal, cutoff, sfreq, order=5):
+    """Apply IIR high-pass Butterworth filter.
+
+    Parameters
+    ----------
+    signal : ndarray
+        The signal to be filtered.
+    cutoff : float
+        Cutoff frequency. Passband is above `cutoff`.
+    sfreq : int
+        Sampling frequency of `signal`.
+    order : int, optional
[…]
