+++ biopeaks/io_utils.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Input/output utilities."""
 
 import json
 import pandas as pd
@@ -9,8 +10,27 @@ from pathlib import Path
 
 
 def read_custom(rpath, customheader, channeltype):
+    """Read a channel from a plain text file.
 
-    # Prepare output.
+    Parameters
+    ----------
+    rpath : str
+        File system location of the file.
+    customheader : dict
+        Dictionary containing header information (channel index,
+        column separator, number of header rows, sampling frequency).
+    channeltype : str
+        The kind of channel to read. One of {"marker", "signal"}.
+
+    Returns
+    -------
+    output : dict
+        Dictionary containing the signal corresponding to the requested
+        channel and any error raised while reading the signal from file. If
+        `channeltype` is "signal", `output` also contains a vector containing
+        the  seconds corresponding to the samples in the signal, as well as the
+        signal's sampling frequency.
+    """
     output = {"error": False,
               "sec": None,
               "signal": None,
@@ -37,8 +57,7 @@ def read_custom(rpath, customheader, channeltype):
     if channeltype == "signal":
         sfreq = customheader["sfreq"]
         signallen = signal.size
-        sec = np.linspace(0, signallen / sfreq, signallen)
-        output["sec"] = sec
+        output["sec"] = np.linspace(0, signallen / sfreq, signallen)
         output["sfreq"] = sfreq
 
     output["signal"] = np.ravel(signal)
@@ -47,23 +66,32 @@ def read_custom(rpath, customheader, channeltype):
 
 
 def write_custom(rpath, wpath, segment, customheader):
+    """Write segmented channels to a plain text file.
+
+    Save a copy of a plain text file after segmenting all channels it contains.
+
+    Parameters
+    ----------
+    rpath : str
+        File system location to read the original file from.
+    wpath : str
+        File system location to write the segmented copy of the file to.
+    segment : list of float
+        Start and end of the user-selected segment in seconds. This segment
+        will be retained for all channels in the copy of the file.
+    customheader : dict
+        Dictionary containing custom header information (column separator,
+        number of header rows, sampling frequency).
     """
-    segment : list
-    Start and end of segments in seconds.
-    """
-    # Get the header.
     with open(rpath, "rt") as oldfile:
-        header = [line for line in islice(oldfile, customheader["skiprows"])]
-    # Get the data.
+        header = [line for line in islice(oldfile, customheader["skiprows"])]    # original header
     data = pd.read_csv(rpath, sep=customheader["separator"], header=None,
-                       skiprows=customheader["skiprows"])
-    # Apply segmentation to all channels.
+                       skiprows=customheader["skiprows"])    # original channels
     begsamp = int(np.rint(segment[0] * customheader["sfreq"]))
     endsamp = int(np.rint(segment[1] * customheader["sfreq"]))
-    data = data.iloc[begsamp:endsamp, :]
+    data = data.iloc[begsamp:endsamp, :]    # apply segmentation to all channels
 
-    # Write header and segmented data to the new file.
-    with open(wpath, "w", newline='') as newfile:
+    with open(wpath, "w", newline='') as newfile:    # write header and segmented channels to new file
         for line in header:
             newfile.write(line)
         data.to_csv(newfile, sep=customheader["separator"], header=False,
@@ -71,8 +99,26 @@ def write_custom(rpath, wpath, segment, customheader):
 
 
 def read_opensignals(rpath, channel, channeltype):
+    """Read a channel from an OpenSignals file.
 
-    # Prepare output.
+    Parameters
+    ----------
[…]
