+++ biopeaks/analysis_utils.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Generic analysis utilities."""
+"""Signal processing utilities used in heart and resp modules."""
 
 import numpy as np
 from scipy.interpolate import interp1d
@@ -15,12 +15,13 @@ def interp_stats(peaks, stats, nsamp):
     ----------
     peaks : ndarray
         Samples associated with the instantaneous statistics. Must have the same
-        number of elements as stats.
+        number of elements as `stats`.
     stats : ndarray
         The instantaneous statistics associated with each peak. Must have the
-        same number of elements as peaks.
+        same number of elements as `peaks`.
     nsamp : int
-        Interpolate statistics over a vector containing samples from 0 to nsamp.
+        Interpolate statistics over a vector containing samples from 0 to
+        `nsamp`.
 
     Returns
     -------
@@ -34,10 +35,10 @@ def interp_stats(peaks, stats, nsamp):
     Notes
     -----
     Stats values for the samples up until the first peak and from last peak to
-    the last sample (nsamp) are extrapolated using the value of the first and
-    last element of stats respectively. First order spline interpolation is
+    the last sample (`nsamp`) are extrapolated using the value of the first and
+    last element of `stats` respectively. First order spline interpolation is
     used, since higher order interpolation (e.g., cubic) can lead to
-    biologically implausible interpolated stats values due to over- or
+    biologically implausible interpolated `stats` values due to over- or
     undershooting (i.e., violations of monotonicity).
     """
     f = interp1d(np.ravel(peaks), stats, kind='slinear',

[…]
