### 2020-10-13 ece4d39
Worked on docstrings.
