+++ biopeaks/heart.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Extract features from cardiac signals."""
 
 import numpy as np
 import pandas as pd
@@ -11,11 +12,46 @@ from .analysis_utils import interp_stats
 
 
 def ecg_peaks(signal, sfreq, smoothwindow=.1, avgwindow=.75,
-              gradthreshweight=1.5, minlenweight=0.4, mindelay=.3,
+              gradthreshweight=1.5, minlenweight=.4, mindelay=.3,
               enable_plot=False):
-    """
-    enable_plot is for debugging and demonstration purposes when the function
-    is called in isolation.
+    """Detect R-peaks in an electrocardiogram (ECG).
+
+    QRS complexes are detected based on the steepness of the absolute gradient
+    of the ECG signal. Subsequently, R-peaks are detected as local maxima in
+    the QRS complexes.
+
+    Parameters
+    ----------
+    signal : ndarray
+        The ECG signal.
+    sfreq : int
+        The sampling frequency of `signal`.
+    smoothwindow : float, optional
+        Size of the kernel used for smoothing the absolute gradient of `signal`.
+        In seconds. Default is .1.
+    avgwindow : float, optional
+        Size of the kernel used for computing the local average of the smoothed
+        absolute gradient of `signal`. In seconds. Default is .75.
+    gradthreshweight : float, optional
+        Factor used to offset the averaged absolute gradient of `signal`. The
+        resulting time series is then used as a threshold for QRS detection.
+        Default is 1.5.
+    minlenweight : float, optional
+        The average QRS duration is multiplied by `minlenweight` in order to
+        obtain a threshold for the minimal duration of QRS complexes (QRS
+        complexes shorted than the minimal duration will be discared). Default
+        is .4.
+    mindelay : float, optional
+        Minimal delay between R-peaks. R-peaks that follow other R-peaks by less
+        than `mindelay` will be discarded. In seconds. Default is .3.
+    enable_plot : bool, optional
+        Visualize `signal` along with the detection thresholds, as well as the
+        detected QRS complexes and R-peaks. Default is False.
+
+    Returns
+    -------
+    peaks : ndarray
+        The samples within `signal` that mark the occurrences of R-peaks.
     """
     if enable_plot:
         plt.figure()
@@ -32,20 +68,16 @@ def ecg_peaks(signal, sfreq, smoothwindow=.1, avgwindow=.75,
     gradthreshold = gradthreshweight * avggrad
     mindelay = int(np.rint(sfreq * mindelay))
 
-    # Visualize thresholds.
     if enable_plot:
         ax1.plot(filt)
         ax2.plot(smoothgrad)
         ax2.plot(gradthreshold)
 
-    # Identify start and end of QRS complexes.
     qrs = smoothgrad > gradthreshold
     beg_qrs = np.where(np.logical_and(np.logical_not(qrs[0:-1]), qrs[1:]))[0]
     end_qrs = np.where(np.logical_and(qrs[0:-1], np.logical_not(qrs[1:])))[0]
-    # Throw out QRS-ends that precede first QRS-start.
-    end_qrs = end_qrs[end_qrs > beg_qrs[0]]
+    end_qrs = end_qrs[end_qrs > beg_qrs[0]]    # discard QRS ends that precede first QRS start
 
-    # Identify R-peaks within QRS (ignore QRS that are too short).
     num_qrs = min(beg_qrs.size, end_qrs.size)
     min_len = np.mean(end_qrs[:num_qrs] - beg_qrs[:num_qrs]) * minlenweight
     peaks = [0]
[…]
