### 2020-10-14 622e20f
Worked on docstrings.
