+++ biopeaks/benchmarks/benchmark_ECG_local.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
+"""Benchmark the heart.ecg_peaks detector.
+
+Before running this script, please download the Glasgow University Database
+(GUDB) from http://researchdata.gla.ac.uk/716/. Then specify `data_dir`,
+`tolerance`, and `condition`.
+"""
 
 import numpy as np
 import pandas as pd
@@ -8,12 +14,13 @@ from wfdb.processing import compare_annotations
 
 
 data_dir = Path(".../experiment_data")    # replace with your local "experiment_data" directory once you've downloaded the database
-
-condition = "sitting"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
-sfreq = 250
 tolerance = 1    # in samples
+condition = "sitting"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
+
+sfreq = 250
 print(f"Setting tolerance for match between algorithmic and manual annotation"
-      f" to 1 sample, corresponding to {1 / sfreq} seconds at a sampling rate of {sfreq}.")
+      f" to {tolerance} sample(s), corresponding to {tolerance / sfreq} seconds"
+      f" at a sampling rate of {sfreq}.")
 
 sensitivity = []
 precision = []
