+++ biopeaks/benchmarks/benchmark_ECG_stream.py
@@ -1,8 +1,236 @@
+# -*- coding: utf-8 -*-
+"""Benchmark the heart.ecg_peaks detector.
+
+Stream the ECG and annotation files from the Glasgow University Database (GUDB),
+hosted at https://berndporr.github.io/ECG-GUDB/experiment_data. You can specify
+`condition`, `channel`, `annotation`, and `tolerance` before running the script.
+"""
+
 from biopeaks.heart import ecg_peaks
-from benchmark_utils import BenchmarkDetectorGUDB
+import asyncio
+from io import StringIO
+from timeit import default_timer as timer
+import aiohttp
+import numpy as np
+from wfdb.processing import compare_annotations
 
-condition = "sitting"    # can be one of {"sitting", "maths", "walking", "hand_bike", "jogging"}
-tolerance = 1    # in samples
 
-pipeline = BenchmarkDetectorGUDB(ecg_peaks, tolerance)
-pipeline.benchmark_records(condition, channel="einthoven_II", annotation="annotation_cables")
+class BenchmarkDetectorGUDB:
+    """Evaluate an ECG R-peak detector on datasets from the GUDB database."""
+
+    channels = {"cs_V2_V1": 0, "einthoven_II": 1, "einthoven_III": 2}
+    base_url = "https://berndporr.github.io/ECG-GUDB/experiment_data/subject_"
+
+    def __init__(self, detector, tolerance, sfreq=250, n_runs=100):
+        """Configure benchmarking for a detector.
+
+        Parameters
+        ----------
+        detector : function
+            A function that takes a vector containing a physiological record
+            as first positional argument and an integer sampling rate as second
+            positional argument. Must return a vector containing the detected
+            extrema.
+        tolerance : int
+            Maximum difference in samples that is permitted between the manual
+            annotation and the annotation generated by the detector.
+        sfreq : int, optional
+            The sampling frequency of the GUDB records. Default is 250.
+        n_runs : int, optional
+            The number of runs used for obtaining the average run time of the
+            detector. The default is 100.
+        """
+        self.detector = detector
+        self.tolerance = tolerance
+        self.sfreq = sfreq
+        self.n_runs = n_runs
+        self.session = None
+        self.queue = None
+        self.channel = None
+        self.annotation = None
+        self.urls = None
+
+    async def score_record(self, record, annotation):
+        """Obtain detector performance for an annotated record.
+
+        Parameters
+        ----------
+        record : ndarray
+            The raw physiological record.
+        annotation : ndarray
+            The manual extrema annotations.
+
+        Returns
+        -------
+        precision : float
+            The detectors precision on the record given the tolerance.
+        sensitivity : float
+            The detectors sensitivity on the record given the tolerance.
+        """
+        detector_annotation = self.detector(record, self.sfreq)
+
+        comparitor = compare_annotations(annotation, detector_annotation,
+                                         self.tolerance)
+        tp = comparitor.tp
+        fp = comparitor.fp
+        fn = comparitor.fn
+        sensitivity = tp / (tp + fn)
+        precision = tp / (tp + fp)
+
+        return precision, sensitivity
+
+    async def time_record(self, record):
+        """Obtain the average run time of a detector on a record over N runs.
+
+        Parameters
+        ----------
+        record : ndarray
+            The raw physiological record.
+
+        Returns
+        -------
+        avg_time : int
+            The run time of the detector on the record averaged over n_runs. In
+            milliseconds.
+        """
+        start = timer()
+
+        for _ in range(self.n_runs):
+            self.detector(record, self.sfreq)
+
+        end = timer()
+        avg_time = (end - start) / self.n_runs * 1000
+
+        return avg_time
+
[…]

[…]
