### 2020-10-14 JanCBrammer/biopeaks#11 @JanCBrammer
@sappelhoff, @TomDonoghue, all modules, classes/methods, and functions are now documented according to [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html) conventions. I agree that this is necessary and useful, despite `biopeaks` mostly containing GUI code.
