### 2020-10-14 JanCBrammer/biopeaks#8 @JanCBrammer
Closing and transfering to #21, since original issue has been addressed.
