### 2020-10-14 #2621 @JanCBrammer
@sappelhoff, @TomDonoghue, I've addressed all "blocking" issues (JanCBrammer/biopeaks#8, JanCBrammer/biopeaks#10, JanCBrammer/biopeaks#11, JanCBrammer/biopeaks#16) and the repository is now ready for your final evaluation.
