### 2020-10-14 JanCBrammer/biopeaks#11 @sappelhoff
great, I see many docstr improvements! Thanks :-)
