### 2020-10-14 #2621 @sappelhoff
I checked all issues that I raised with @JanCBrammer and everything has been addressed to my full satisfaction. `biopeaks` is a useful piece of software that will make life easier for its users. The accompanying paper is well written, the documentation is up to date an thorough. Unit tests and coverage reports are in place, and best practices of open source software (and open science, where applicable) are being followed. I see a need for software like `biopeaks` that provides a GUI for users that are not (yet) familiar with scripting analyses.

I recommend accepting this project for a publication in JOSS.
