### 2020-10-16 cde8c03
Pull out dedicated funtion to ensure peak-trough alternation.
