### 2020-10-23 42b5d54
Consider extrema with equal amplitude while ensuring peak-trough alternation.
