### 2020-10-23 598844e
Unit-test all functions in heart and resp modules (in addition to functional tests in test_gui).
