### 2020-10-23 0e47feb
Accomodate system-specific rounding shenanigans in assertions.
