### 2020-10-24 d46e3b1
Get coverage for methods executed in QThreadPool.
