### 2020-10-24 07f0af4
Read data in fixtures.
