### 2020-10-24 fcfdd7b
Fixing linter complaints.
