### 2020-10-25 #2621 @jamesheathers
Hello all,

My apologies for the delay. There was also another problem: I can't find anything to change here.

* I got this to run cleanly and straightforwardly with a fresh install of Miniconda
* All of the core functionality I would expect for a R-peak/systolic interval detector is present, and it all works to a higher reliability than simpler peak detection methods
* I gave it an engineer (not Python naive but biosignal analysis naive) and he was immediately able to get the software to function as intended and to my satisfaction
* Any additional suggestions for analytic functions I could make would go beyond the scope of the paper
* The documentation is extremely good

In short, I can't actually find anything I'd want to change that makes a reasonable contribution here. I don't think I've ever written an exclusively positive review before, but here we are. This will form a useful tool for beginner to intermediate biosignal analysts who wish to explore cardiac cyclic time series.
