### 2020-10-26 7a5836e
Point PyPI project description to GitHub.
