### 2020-10-26 #2621 @JanCBrammer
@oliviaguest, I released a new version ([v.1.4.1](https://github.com/JanCBrammer/biopeaks/releases/tag/v1.4.1)) containing the improvements related to this review. I've then deposited the repository at [10.5281/zenodo.4133324](http://doi.org/10.5281/zenodo.4133324).
