### 2020-10-26 JanCBrammer/biopeaks#7 @JanCBrammer
Closing this after updating `setup.py` accordingly.
